+++ postgres/datadog_checks/postgres/postgres.py
@@ -16,7 +16,7 @@ from datadog_checks.base.utils.db.utils import (
 )
 from datadog_checks.base.utils.db.utils import resolve_db_host as agent_host_resolver
 from datadog_checks.base.utils.serialization import json
-from datadog_checks.postgres import aws
+from datadog_checks.postgres import aws, azure
 from datadog_checks.postgres.connections import MultiDatabaseConnectionPool
 from datadog_checks.postgres.discovery import PostgresAutodiscovery
 from datadog_checks.postgres.metadata import PostgresMetadata
@@ -614,6 +614,10 @@ class PostgreSql(AgentCheck):
                     port=self._config.port,
                     region=region,
                 )
+            client_id = self._config.managed_identity.get('client_id', None)
+            scope = self._config.managed_identity.get('identity_scope', None)
+            if client_id is not None:
+                password = azure.generate_managed_identity_token(client_id=client_id, scope=scope)
 
             args = {
                 'host': self._config.host,