+++ postgres/CHANGELOG.md
@@ -5,6 +5,8 @@
 ***Added***:
 
 * Add support for sending `database_instance` metadata ([#15559](https://github.com/DataDog/integrations-core/pull/15559))
+* Update dependencies for Agent 7.48 ([#15585](https://github.com/DataDog/integrations-core/pull/15585))
+* Add support for authenticating through Azure Managed Identity ([#15609](https://github.com/DataDog/integrations-core/pull/15609))
 
 ## 14.1.0 / 2023-08-10